Added another chart type to draw IQ graph with trajectory trace. also update c++ standard to 17

+++ chart/src/GridShape.cpp
@@ -339,7 +339,7 @@ namespace Presentation
                 float zStart = kMinZValue;
                 float zStop = kMaxZValue;
                 float xValue = (float)GridRectangle.X;
-                if(_showRightPlane) xValue += GridRectangle.Width;
+                if(_showRightPlane) xValue += (float)GridRectangle.Width;
 
                 sf::Color gridColor = ColorUtil::ColorFrom(GridColor);
